@@ -22,13 +22,12 @@ import {MedicalCareNewMedicalReportMedicinePage} from '../medical-care-new-medic
 import {PassDataService} from '../../services/pass-data.service';
 
 @Component({
-  selector: 'app-medical-care-new-medical-report',
-  templateUrl: './medical-care-new-medical-report.page.html',
-  styleUrls: ['./medical-care-new-medical-report.page.scss'],
+  selector: "app-medical-care-new-medical-report",
+  templateUrl: "./medical-care-new-medical-report.page.html",
+  styleUrls: ["./medical-care-new-medical-report.page.scss"],
 })
 export class MedicalCareNewMedicalReportPage implements OnInit {
-
-  @ViewChild('ionImgSlides') slides: IonSlides;
+  @ViewChild("ionImgSlides", { static: false }) slides: IonSlides;
   todayDate;
   pageNames;
   operations;
@@ -47,7 +46,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   incidentShowDate;
   checkupShowDate;
   medicalRecords: MedicalRecord = new MedicalRecord();
-  prescription = {'id': null, 'medicalRecords': null, 'medications': []};
+  prescription = { id: null, medicalRecords: null, medications: [] };
   fullMedicalReport: any;
   incident;
   checkup;
@@ -58,7 +57,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   incidentAnswer: any[] = [];
   incidentAnswersNoOfItems: any[] = [];
   incidentQuestionsEditParamTemps: any[] = [];
-  @ViewChild('file') inputEl: ElementRef;
+  @ViewChild("file", { static: false }) inputEl: ElementRef;
   /////Templet Data CHECKUP///////
   checkupTemplate: any;
   checkupQuestions: any;
@@ -67,8 +66,81 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   checkupAnswersNoOfItems: any[] = [];
   checkupQuestionsEditParamTemps: any[] = [];
   loopData;
-  showAllTimes = [{'id': '', 'time': '00:00'}, {'id': '', 'time': '00:30'}, {'id': '', 'time': '01:00'}, {'id': '', 'time': '01:30'}, {'id': '', 'time': '02:00'}, {'id': '', 'time': '02:30'}, {'id': '', 'time': '03:00'}, {'id': '', 'time': '03:30'}, {'id': '', 'time': '04:00'}, {'id': '', 'time': '04:30'}, {'id': '', 'time': '05:00'}, {'id': '', 'time': '05:30'}, {'id': '', 'time': '06:00'}, {'id': '', 'time': '06:30'}, {'id': '', 'time': '07:00'}, {'id': '', 'time': '07:30'}, {'id': '', 'time': '08:00'}, {'id': '', 'time': '08:30'}, {'id': '', 'time': '09:00'}, {'id': '', 'time': '09:30'}, {'id': '', 'time': '10:00'}, {'id': '', 'time': '10:30'}, {'id': '', 'time': '11:00'}, {'id': '', 'time': '11:30'}, {'id': '', 'time': '12:00'}, {'id': '', 'time': '12:30'}, {'id': '', 'time': '13:00'}, {'id': '', 'time': '13:30'}, {'id': '', 'time': '14:00'}, {'id': '', 'time': '14:30'}, {'id': '', 'time': '15:00'}, {'id': '', 'time': '15:30'}, {'id': '', 'time': '16:00'}, {'id': '', 'time': '16:30'}, {'id': '', 'time': '17:00'}, {'id': '', 'time': '17:30'}, {'id': '', 'time': '18:00'}, {'id': '', 'time': '18:30'}, {'id': '', 'time': '19:00'}, {'id': '', 'time': '19:30'}, {'id': '', 'time': '20:00'}, {'id': '', 'time': '20:30'}, {'id': '', 'time': '21:00'}, {'id': '', 'time': '21:30'}, {'id': '', 'time': '22:00'}, {'id': '', 'time': '22:30'}, {'id': '', 'time': '23:00'}, {'id': '', 'time': '23:30'}];
-  fileTypes = ["jpg", "jpeg", "png", "gif", "ico", "bmp", "webp", "tiff", "pdf", "txt", "xls", "xlsx", "doc", "docx", "ppt", "pptx", "mp4", "flv", "avi", "mov", "wmv", "mp3", "wma"];
+  showAllTimes = [
+    { id: "", time: "00:00" },
+    { id: "", time: "00:30" },
+    { id: "", time: "01:00" },
+    { id: "", time: "01:30" },
+    { id: "", time: "02:00" },
+    { id: "", time: "02:30" },
+    { id: "", time: "03:00" },
+    { id: "", time: "03:30" },
+    { id: "", time: "04:00" },
+    { id: "", time: "04:30" },
+    { id: "", time: "05:00" },
+    { id: "", time: "05:30" },
+    { id: "", time: "06:00" },
+    { id: "", time: "06:30" },
+    { id: "", time: "07:00" },
+    { id: "", time: "07:30" },
+    { id: "", time: "08:00" },
+    { id: "", time: "08:30" },
+    { id: "", time: "09:00" },
+    { id: "", time: "09:30" },
+    { id: "", time: "10:00" },
+    { id: "", time: "10:30" },
+    { id: "", time: "11:00" },
+    { id: "", time: "11:30" },
+    { id: "", time: "12:00" },
+    { id: "", time: "12:30" },
+    { id: "", time: "13:00" },
+    { id: "", time: "13:30" },
+    { id: "", time: "14:00" },
+    { id: "", time: "14:30" },
+    { id: "", time: "15:00" },
+    { id: "", time: "15:30" },
+    { id: "", time: "16:00" },
+    { id: "", time: "16:30" },
+    { id: "", time: "17:00" },
+    { id: "", time: "17:30" },
+    { id: "", time: "18:00" },
+    { id: "", time: "18:30" },
+    { id: "", time: "19:00" },
+    { id: "", time: "19:30" },
+    { id: "", time: "20:00" },
+    { id: "", time: "20:30" },
+    { id: "", time: "21:00" },
+    { id: "", time: "21:30" },
+    { id: "", time: "22:00" },
+    { id: "", time: "22:30" },
+    { id: "", time: "23:00" },
+    { id: "", time: "23:30" },
+  ];
+  fileTypes = [
+    "jpg",
+    "jpeg",
+    "png",
+    "gif",
+    "ico",
+    "bmp",
+    "webp",
+    "tiff",
+    "pdf",
+    "txt",
+    "xls",
+    "xlsx",
+    "doc",
+    "docx",
+    "ppt",
+    "pptx",
+    "mp4",
+    "flv",
+    "avi",
+    "mov",
+    "wmv",
+    "mp3",
+    "wma",
+  ];
   MULTI_SHORT_TEXT_ONE_VIEW_SELECTED_Index;
   DROPDOWN_MENU_ONE_VIEW_SELECTED_index;
   SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED_INDEX;
@@ -81,7 +153,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   checkupSelectedDate;
   selectedIncidentTime;
   selectedCheckupTime;
-  phoneNumber: string = '';
+  phoneNumber: string = "";
   //////////////EDIT VIEW///////////////
   EditView = false;
   tempFullMedicalRecord;
@@ -93,16 +165,25 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   // @Input() selectedIncidentDate:any;
   // @Input() medicalRecord:any;
 
-  constructor(private config: NgSelectConfig, private modalCtrl: ModalController,public passData:PassDataService,
-              private medicalService: MedicalCareService, private alrtCtrl: AlertController, private loadCtrl: LoadingViewService,
-              private classServ: ClassesService, private studentServ: StudentsService, private transDate: TransFormDateService,
-              private accountServ: AccountService, private checkboxFunctionService: CheckboxFunctionService,
-              public network: Network,private compress:ImageCompressorService,private toastCtrl:ToastViewService,
-              private bkgInc:BackgroundMedicalcareService, private platform:Platform)
-  {
-    this.config.notFoundText = 'No match found';
-
-
+  constructor(
+    private config: NgSelectConfig,
+    private modalCtrl: ModalController,
+    public passData: PassDataService,
+    private medicalService: MedicalCareService,
+    private alrtCtrl: AlertController,
+    private loadCtrl: LoadingViewService,
+    private classServ: ClassesService,
+    private studentServ: StudentsService,
+    private transDate: TransFormDateService,
+    private accountServ: AccountService,
+    private checkboxFunctionService: CheckboxFunctionService,
+    public network: Network,
+    private compress: ImageCompressorService,
+    private toastCtrl: ToastViewService,
+    private bkgInc: BackgroundMedicalcareService,
+    private platform: Platform
+  ) {
+    this.config.notFoundText = "No match found";
 
     this.todayDate = this.transDate.transformTheDate(new Date(), "dd-MM-yyyy");
     this.operations = this.passData.dataToPass.operation;
@@ -110,7 +191,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     if (this.passData.dataToPass.for) {
       this.pageNames = "New " + this.passData.dataToPass.for;
       this.addCheckup = false;
-      if(this.operations == "edit"){
+      if (this.operations == "edit") {
         this.EditView = true;
         this.pageNames = this.passData.dataToPass.pageName;
       }
@@ -119,38 +200,54 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         this.incidentMinDate = new Date(2016, 0, 1);
         this.incidentMaxDate = new Date();
         this.incidentShowDate = new FormControl(new Date()).value;
-        this.selectedIncidentTime = {'id': '', 'time': '00:00'};
-        this.incidentSelectedDate = this.transDate.transformTheDate(new Date(), "dd-MM-yyyy");
+        this.selectedIncidentTime = { id: "", time: "00:00" };
+        this.incidentSelectedDate = this.transDate.transformTheDate(
+          new Date(),
+          "dd-MM-yyyy"
+        );
         this.incidentAnswer[0] = {
-          "incidentAnswersObjectsList": []
+          incidentAnswersObjectsList: [],
         };
-        if(this.operations == "new") {
+        if (this.operations == "new") {
           this.getIncidentTemplate();
         }
       } else {
         this.checkupMinDate = new Date(2016, 0, 1);
         this.checkupMaxDate = new Date();
         this.checkupShowDate = new FormControl(new Date()).value;
-        this.selectedCheckupTime = {'id': '', 'time': '00:00'};
-        this.checkupSelectedDate = this.transDate.transformTheDate(new Date(), "dd-MM-yyyy");
+        this.selectedCheckupTime = { id: "", time: "00:00" };
+        this.checkupSelectedDate = this.transDate.transformTheDate(
+          new Date(),
+          "dd-MM-yyyy"
+        );
         this.checkupAnswer = {
-          "checkupAnswersObjectsList": []
+          checkupAnswersObjectsList: [],
         };
-        if(this.operations == "new") {
+        if (this.operations == "new") {
           this.getCheckUpTemplate();
         }
       }
-    } else if(this.passData.dataToPass.forAdd && this.passData.dataToPass.operation == 'new'){
+    } else if (
+      this.passData.dataToPass.forAdd &&
+      this.passData.dataToPass.operation == "new"
+    ) {
       let date = this.passData.dataToPass.selectedIncidentDate;
       this.pageNames = "Add " + this.passData.dataToPass.forAdd;
       this.addCheckup = true;
-      this.checkupMinDate = new Date(parseInt(date[2]), parseInt(date[1])-1, parseInt(date[0]));
+      this.checkupMinDate = new Date(
+        parseInt(date[2]),
+        parseInt(date[1]) - 1,
+        parseInt(date[0])
+      );
       this.checkupMaxDate = new Date();
       this.checkupShowDate = new FormControl(new Date()).value;
-      this.selectedCheckupTime = {'id': '', 'time': '00:00'};
-      this.checkupSelectedDate = this.transDate.transformTheDate(new Date(), "dd-MM-yyyy");
+      this.selectedCheckupTime = { id: "", time: "00:00" };
+      this.checkupSelectedDate = this.transDate.transformTheDate(
+        new Date(),
+        "dd-MM-yyyy"
+      );
       this.checkupAnswer = {
-        "checkupAnswersObjectsList": []
+        checkupAnswersObjectsList: [],
       };
       this.getCheckUpTemplate();
     }
@@ -162,65 +259,91 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
       this.getAllStudents();
     }
 
-
-    if(this.operations == "edit"){
+    if (this.operations == "edit") {
       this.EditView = true;
       this.tempFullMedicalRecord = this.passData.dataToPass.medicalRecord;
-      this.setSelectedClassFromEdit(this.tempFullMedicalRecord.medicalRecord.student.classes);
-      this.setSelectedStudentFromEdit(this.tempFullMedicalRecord.medicalRecord.student);
-      this.medicalRecords=this.tempFullMedicalRecord.medicalRecord;
+      this.setSelectedClassFromEdit(
+        this.tempFullMedicalRecord.medicalRecord.student.classes
+      );
+      this.setSelectedStudentFromEdit(
+        this.tempFullMedicalRecord.medicalRecord.student
+      );
+      this.medicalRecords = this.tempFullMedicalRecord.medicalRecord;
       if (this.passData.dataToPass.for == "Incident") {
         this.newIncident = true;
         this.incidentTitle = this.tempFullMedicalRecord.medicalRecord.incident.title;
-        if(this.tempFullMedicalRecord.medicalRecord.incident.followUpPhone) {
+        if (this.tempFullMedicalRecord.medicalRecord.incident.followUpPhone) {
           this.phoneNumber = this.tempFullMedicalRecord.medicalRecord.incident.followUpPhone.toString();
         }
-        let Date_time_Array = this.tempFullMedicalRecord.medicalRecord.incident.incidentDate.split(" ");
+        let Date_time_Array = this.tempFullMedicalRecord.medicalRecord.incident.incidentDate.split(
+          " "
+        );
         let DateArray = Date_time_Array[0].split("-");
-        this.incidentSelectedDate =Date_time_Array[0];
+        this.incidentSelectedDate = Date_time_Array[0];
         this.prescription = this.tempFullMedicalRecord.medicalRecord.prescription;
-        this.incidentShowDate =new Date(parseInt(DateArray[2]),parseInt(DateArray[1])-1,parseInt(DateArray[0]));
-        this.selectedIncidentTime = {'id': '', 'time': Date_time_Array[1]};
+        this.incidentShowDate = new Date(
+          parseInt(DateArray[2]),
+          parseInt(DateArray[1]) - 1,
+          parseInt(DateArray[0])
+        );
+        this.selectedIncidentTime = { id: "", time: Date_time_Array[1] };
         this.incident = this.tempFullMedicalRecord.medicalRecord.incident;
         this.incident.answers = this.tempFullMedicalRecord.incidentAnswers;
-        if(this.tempFullMedicalRecord.medicalRecord.checkup){
+        if (this.tempFullMedicalRecord.medicalRecord.checkup) {
           this.checkup = this.tempFullMedicalRecord.medicalRecord.checkup;
         }
-        this.setIncidentTempletFromEdit(this.tempFullMedicalRecord.incidentTemplate);
-      }else{
+        this.setIncidentTempletFromEdit(
+          this.tempFullMedicalRecord.incidentTemplate
+        );
+      } else {
         this.checkupTitle = this.tempFullMedicalRecord.medicalRecord.checkup.title;
         this.prescription = this.tempFullMedicalRecord.medicalRecord.prescription;
-        let Date_time_Array = this.tempFullMedicalRecord.medicalRecord.checkup.checkupDate.split(" ");
+        let Date_time_Array = this.tempFullMedicalRecord.medicalRecord.checkup.checkupDate.split(
+          " "
+        );
         let DateArray = Date_time_Array[0].split("-");
-        this.checkupSelectedDate =Date_time_Array[0];
-        this.checkupShowDate =new Date(parseInt(DateArray[2]),parseInt(DateArray[1])-1,parseInt(DateArray[0]));
-        this.selectedCheckupTime = {'id': '', 'time': Date_time_Array[1]};
+        this.checkupSelectedDate = Date_time_Array[0];
+        this.checkupShowDate = new Date(
+          parseInt(DateArray[2]),
+          parseInt(DateArray[1]) - 1,
+          parseInt(DateArray[0])
+        );
+        this.selectedCheckupTime = { id: "", time: Date_time_Array[1] };
         this.checkup = this.tempFullMedicalRecord.medicalRecord.checkup;
         this.checkup.answers = this.tempFullMedicalRecord.checkupAnswers;
-        this.setCheckupTempletFromEdit(this.tempFullMedicalRecord.checkupTemplate);
+        this.setCheckupTempletFromEdit(
+          this.tempFullMedicalRecord.checkupTemplate
+        );
       }
     }
 
-    if(this.passData.dataToPass.forAdd && this.passData.dataToPass.operation == 'edit') {
+    if (
+      this.passData.dataToPass.forAdd &&
+      this.passData.dataToPass.operation == "edit"
+    ) {
       let date = this.passData.dataToPass.selectedIncidentDate;
       this.pageNames = "Add " + this.passData.dataToPass.forAdd;
       this.addCheckup = true;
-      this.checkupMinDate = new Date(parseInt(date[2]), parseInt(date[1])-1, parseInt(date[0]));
+      this.checkupMinDate = new Date(
+        parseInt(date[2]),
+        parseInt(date[1]) - 1,
+        parseInt(date[0])
+      );
       this.checkupMaxDate = new Date();
       this.checkupShowDate = new FormControl(new Date()).value;
-      this.selectedCheckupTime = {'id': '', 'time': '00:00'};
-      this.checkupSelectedDate = this.transDate.transformTheDate(new Date(), "dd-MM-yyyy");
+      this.selectedCheckupTime = { id: "", time: "00:00" };
+      this.checkupSelectedDate = this.transDate.transformTheDate(
+        new Date(),
+        "dd-MM-yyyy"
+      );
       this.checkupAnswer = {
-        "checkupAnswersObjectsList": []
+        checkupAnswersObjectsList: [],
       };
       this.getCheckUpTemplate();
     }
-
-
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   close() {
     this.modalCtrl.dismiss();
@@ -236,10 +359,15 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     } else {
       if (this.newIncident) {
         let phoneNumberFound = false;
-        if(this.phoneNumber.length < 1 || this.phoneNumber.length > 7){
+        if (this.phoneNumber.length < 1 || this.phoneNumber.length > 7) {
           phoneNumberFound = true;
         }
-        if (this.incidentTitle && this.selectedStudent && this.selectedClass && phoneNumberFound) {
+        if (
+          this.incidentTitle &&
+          this.selectedStudent &&
+          this.selectedClass &&
+          phoneNumberFound
+        ) {
           return false;
         } else {
           return true;
@@ -254,66 +382,65 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     }
   }
 
-
-  async fabSelected(index, fab:IonFab) {
-    let dateArray:any[];
-    if(this.incidentSelectedDate){
+  async fabSelected(index, fab: IonFab) {
+    let dateArray: any[];
+    if (this.incidentSelectedDate) {
       dateArray = this.incidentSelectedDate.split("-");
-    }else{
+    } else {
       dateArray = this.checkupSelectedDate.split("-");
     }
 
     await fab.close();
     let modal;
     if (index == 0) {
-      let data = {Date: this.todayDate, operation: 'new'};
+      let data = { Date: this.todayDate, operation: "new" };
 
       this.passData.dataToPass = data;
 
       modal = await this.modalCtrl.create({
         component: MedicalCareNewMedicalReportMedicinePage,
-      componentProps: data});
-
+        componentProps: data,
+      });
     } else {
-
       let data = {
         forAdd: "Checkup",
         Date: this.todayDate,
         operation: this.operations,
-        selectedIncidentDate:dateArray
+        selectedIncidentDate: dateArray,
       };
 
       this.passData.dataToPass = data;
 
       modal = await this.modalCtrl.create({
         component: MedicalCareNewMedicalReportPage,
-      componentProps: data
-    });
-
+        componentProps: data,
+      });
     }
-    modal.onDidDismiss().then(
-        val => {
-          let data = val.data;
-          console.log(data);
-          if (data) {
-            if (data.medication) {
-              if(this.prescription) {
-                this.prescription.medications.push(data.medication);
-              }else{
-                this.prescription = {'id': null, 'medicalRecords': null, 'medications': []};
-                this.prescription.medications.push(data.medication);
-              }
-            } else if (data.checkup) {
-              this.checkup = data.checkup;
-              this.checkupAnswer = data.checkupAnswers;
-              this.checkupTemplate = data.checkupTemplate;
-              this.checkupQuestions = data.checkupTemplate.questionsList;
-            }
+    modal.onDidDismiss().then((val) => {
+      let data = val.data;
+      console.log(data);
+      if (data) {
+        if (data.medication) {
+          if (this.prescription) {
+            this.prescription.medications.push(data.medication);
+          } else {
+            this.prescription = {
+              id: null,
+              medicalRecords: null,
+              medications: [],
+            };
+            this.prescription.medications.push(data.medication);
           }
-        });
+        } else if (data.checkup) {
+          this.checkup = data.checkup;
+          this.checkupAnswer = data.checkupAnswers;
+          this.checkupTemplate = data.checkupTemplate;
+          this.checkupQuestions = data.checkupTemplate.questionsList;
+        }
+      }
+    });
     await modal.present();
   }
-
 
   setStudents() {
     this.studentsLoading = true;
@@ -330,45 +457,86 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
   }
 
   setupMedicalRecord() {
-    this.fullMedicalReport = {'checkupAnswers': [], 'incidentAnswers': [], 'medicalRecord': null};
-    this.incident = {
-      'attachmentsList': [],
-      'followUpPhone': "",
-      'id': null,
-      'details': "",
-      'incidentDate': this.transDate.transformTheDate(new Date(), "dd-MM-yyyy"),
-      'title': "",
-      'answer': []
+    this.fullMedicalReport = {
+      checkupAnswers: [],
+      incidentAnswers: [],
+      medicalRecord: null,
     };
-    this.checkup = {'id': null, "title": '', "details": '', 'checkupTemplate': null, "checkupDate": this.transDate.transformTheDate(new Date(), "dd-MM-yyyy")};
+    this.incident = {
+      attachmentsList: [],
+      followUpPhone: "",
+      id: null,
+      details: "",
+      incidentDate: this.transDate.transformTheDate(new Date(), "dd-MM-yyyy"),
+      title: "",
+      answer: [],
+    };
+    this.checkup = {
+      id: null,
+      title: "",
+      details: "",
+      checkupTemplate: null,
+      checkupDate: this.transDate.transformTheDate(new Date(), "dd-MM-yyyy"),
+    };
   }
 
   allDataDone() {
     if (this.addCheckup) {
       this.checkup.title = this.checkupTitle;
-      this.checkup.checkupDate = this.checkupSelectedDate + " " + this.selectedCheckupTime.time;
+      this.checkup.checkupDate =
+        this.checkupSelectedDate + " " + this.selectedCheckupTime.time;
       this.checkup.checkupTemplate = this.checkupTemplate.id;
       this.modalCtrl.dismiss({
         checkup: this.checkup,
         checkupAnswers: this.checkupAnswer,
-        checkupTemplate: this.checkupTemplate
+        checkupTemplate: this.checkupTemplate,
       });
     } else {
       if (this.newIncident) {
         this.incident.followUpPhone = this.phoneNumber;
-        this.incident.incidentDate = this.incidentSelectedDate + " " + this.selectedIncidentTime.time;
+        this.incident.incidentDate =
+          this.incidentSelectedDate + " " + this.selectedIncidentTime.time;
         this.incident.incidentTemplate = this.incidentTemplate.id;
         this.incident.title = this.incidentTitle;
-        this.bkgInc.addIncident(this.checkup,this.checkupAnswer,this.incident,this.incidentSelectedDate,this.selectedIncidentTime,
-            this.operations,this.checkupTemplate,this.incidentTemplate,this.prescription,this.medicalRecords,this.selectedStudent,
-            this.modalCtrl,this.incidentQuestions,this.incidentAnswer,this.checkupQuestions);
+        this.bkgInc.addIncident(
+          this.checkup,
+          this.checkupAnswer,
+          this.incident,
+          this.incidentSelectedDate,
+          this.selectedIncidentTime,
+          this.operations,
+          this.checkupTemplate,
+          this.incidentTemplate,
+          this.prescription,
+          this.medicalRecords,
+          this.selectedStudent,
+          this.modalCtrl,
+          this.incidentQuestions,
+          this.incidentAnswer,
+          this.checkupQuestions
+        );
       } else {
         this.checkup.title = this.checkupTitle;
-        this.checkup.checkupDate = this.checkupSelectedDate + " " + this.selectedCheckupTime.time;
+        this.checkup.checkupDate =
+          this.checkupSelectedDate + " " + this.selectedCheckupTime.time;
         this.checkup.checkupTemplate = this.checkupTemplate.id;
-        this.bkgInc.addCheckup(this.checkup,this.checkupAnswer,this.incident,this.incidentSelectedDate,this.selectedIncidentTime,
-            this.operations,this.checkupTemplate,this.incidentTemplate,this.prescription,this.medicalRecords,this.selectedStudent,
-            this.modalCtrl,this.incidentQuestions,this.incidentAnswer,this.checkupQuestions);
+        this.bkgInc.addCheckup(
+          this.checkup,
+          this.checkupAnswer,
+          this.incident,
+          this.incidentSelectedDate,
+          this.selectedIncidentTime,
+          this.operations,
+          this.checkupTemplate,
+          this.incidentTemplate,
+          this.prescription,
+          this.medicalRecords,
+          this.selectedStudent,
+          this.modalCtrl,
+          this.incidentQuestions,
+          this.incidentAnswer,
+          this.checkupQuestions
+        );
         // this.medicalRecords.checkup = this.checkup;
       }
       // this.medicalRecords.student = {'id': this.selectedStudent.id};
@@ -381,30 +549,38 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
 
   setSelectedDate(from, ev) {
     if (this.newIncident) {
-      this.incidentSelectedDate = this.transDate.transformTheDate(ev.value, "dd-MM-yyyy");
+      this.incidentSelectedDate = this.transDate.transformTheDate(
+        ev.value,
+        "dd-MM-yyyy"
+      );
     } else {
-      this.checkupSelectedDate = this.transDate.transformTheDate(ev.value, "dd-MM-yyyy");
+      this.checkupSelectedDate = this.transDate.transformTheDate(
+        ev.value,
+        "dd-MM-yyyy"
+      );
     }
   }
 
-  async deleteAttach(attachIndex: any,qNumber) {
+  async deleteAttach(attachIndex: any, qNumber) {
     const alert = await this.alrtCtrl.create({
-      header: 'Alert',
-      message: 'Are you sure that you want to delete this attachment?',
+      header: "Alert",
+      message: "Are you sure that you want to delete this attachment?",
       buttons: [
         {
-          text: 'Cancel',
-          role: 'cancel'
+          text: "Cancel",
+          role: "cancel",
         },
         {
-          text: 'Delete',
-          role: 'destructive',
+          text: "Delete",
+          role: "destructive",
           handler: () => {
-            this.incidentAnswer[0].incidentAnswersObjectsList[qNumber].answer.splice(attachIndex, 1);
+            this.incidentAnswer[0].incidentAnswersObjectsList[
+              qNumber
+            ].answer.splice(attachIndex, 1);
             this.slides.update();
-          }
-        }
-      ]
+          },
+        },
+      ],
     });
     return await alert.present();
   }
@@ -413,14 +589,22 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     if (from == "Medication") {
       this.prescription.medications.splice(index, 1);
     } else if (from == "Checkup") {
-      this.checkup = {'id': null, "title": '', "details": '', 'checkupTemplate': null, "checkupDate": this.transDate.transformTheDate(new Date(), "dd-MM-yyyy")};
+      this.checkup = {
+        id: null,
+        title: "",
+        details: "",
+        checkupTemplate: null,
+        checkupDate: this.transDate.transformTheDate(new Date(), "dd-MM-yyyy"),
+      };
     }
   }
 
-////////////////////////////////////////////GET CLASSES AND STUDENTS
+  ////////////////////////////////////////////GET CLASSES AND STUDENTS
   getAllClasses() {
-    this.classServ.getClassList("Medical Care", 2, null, null, null, null).subscribe(
-        classVal => {
+    this.classServ
+      .getClassList("Medical Care", 2, null, null, null, null)
+      .subscribe(
+        (classVal) => {
           let allData: any = classVal;
 
           // if(this.platform.is('cordova')){
@@ -441,234 +625,276 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
             this.allclasses.push(item);
           }
           this.classesLoading = false;
-        }, async classErr => {
+        },
+        async (classErr) => {
           this.classesLoading = false;
           const alrt = await this.alrtCtrl.create({
-            header: 'Error',
-            subHeader: 'Something went wrong, please refresh the page',
-            buttons: ['OK']
+            header: "Error",
+            subHeader: "Something went wrong, please refresh the page",
+            buttons: ["OK"],
           });
           return alrt.present();
-        });
+        }
+      );
   }
 
   getAllStudents() {
     this.studentServ.getAllStudents(7, "Medical Care").subscribe(
-        studentVal => {
-          let data: any = studentVal;
+      (studentVal) => {
+        let data: any = studentVal;
 
-          // if(this.platform.is('cordova')){
-          //   data = JSON.parse(studentVal.data);
-          // }
+        // if(this.platform.is('cordova')){
+        //   data = JSON.parse(studentVal.data);
+        // }
 
-          this.allStudents = [];
-          for (let value of data) {
-            let students = new Student();
+        this.allStudents = [];
+        for (let value of data) {
+          let students = new Student();
 
-            students.classes.id = value.classes.id;
-            students.classes.name = value.classes.name;
-            students.classes.grade.id = value.classes.grade.id;
-            students.classes.grade.name = value.classes.grade.name;
-            students.classes.branch.id = value.classes.branch.id;
-            students.classes.branch.name = value.classes.branch.name;
-            students.classes.branch.managerId = value.classes.branch.managerId;
-            students.id = value.id;
-            students.name = value.name;
-            students.address = value.address;
-            students.searchByClassGrade = value.classes.grade.name + " - " + value.classes.name;
+          students.classes.id = value.classes.id;
+          students.classes.name = value.classes.name;
+          students.classes.grade.id = value.classes.grade.id;
+          students.classes.grade.name = value.classes.grade.name;
+          students.classes.branch.id = value.classes.branch.id;
+          students.classes.branch.name = value.classes.branch.name;
+          students.classes.branch.managerId = value.classes.branch.managerId;
+          students.id = value.id;
+          students.name = value.name;
+          students.address = value.address;
+          students.searchByClassGrade =
+            value.classes.grade.name + " - " + value.classes.name;
 
-            this.allStudents.push(students);
-          }
-          this.studentsLoading = false;
-        }, async studentErr => {
-          this.studentsLoading = false;
-          const alrt = await this.alrtCtrl.create({
-            header: 'Error',
-            subHeader: 'Something went wrong, please refresh the page',
-            buttons: ['OK']
-          });
-
-          return await alrt.present();
+          this.allStudents.push(students);
+        }
+        this.studentsLoading = false;
+      },
+      async (studentErr) => {
+        this.studentsLoading = false;
+        const alrt = await this.alrtCtrl.create({
+          header: "Error",
+          subHeader: "Something went wrong, please refresh the page",
+          buttons: ["OK"],
         });
+
+        return await alrt.present();
+      }
+    );
   }
 
   ////////////////////////////////////////////////////////////////////////Get The Template Data
   getIncidentTemplate() {
     this.medicalService.getIncidentTemplate().subscribe(
-        val => {
-          this.templateLoading = false;
-          this.incidentTemplate = val;
-          // if(this.platform.is('cordova')){
-          //   this.incidentTemplate = JSON.parse(val.data);
-          // }
+      (val) => {
+        this.templateLoading = false;
+        this.incidentTemplate = val;
+        // if(this.platform.is('cordova')){
+        //   this.incidentTemplate = JSON.parse(val.data);
+        // }
 
-          this.incidentQuestions[0] = this.incidentTemplate.questionsList;
+        this.incidentQuestions[0] = this.incidentTemplate.questionsList;
 
-          this.incidentQuestionsFirst = this.incidentTemplate.questionsList;
+        this.incidentQuestionsFirst = this.incidentTemplate.questionsList;
 
-          for (let i = 0; i < this.incidentQuestionsFirst.length; i++) {
-            this.incidentQuestionsFirst[i].questionNumber = i;
-            this.incidentAnswer[0].incidentAnswersObjectsList[i] = {
-              answer: null
-            };
-            this.incidentAnswersNoOfItems[i] = {
-              noOfItems: null
-            };
-            this.incidentQuestionsFirst[i].editQuestion = false;
-            this.incidentQuestionsFirst[i].isEdited = false;
+        for (let i = 0; i < this.incidentQuestionsFirst.length; i++) {
+          this.incidentQuestionsFirst[i].questionNumber = i;
+          this.incidentAnswer[0].incidentAnswersObjectsList[i] = {
+            answer: null,
+          };
+          this.incidentAnswersNoOfItems[i] = {
+            noOfItems: null,
+          };
+          this.incidentQuestionsFirst[i].editQuestion = false;
+          this.incidentQuestionsFirst[i].isEdited = false;
+        }
+
+        this.incidentQuestions[0] = this.incidentQuestionsFirst;
+
+        for (let i = 0; i < this.incidentQuestions[0].length; i++) {
+          this.mappingDefaultAnswers(
+            this.incidentAnswer[0].incidentAnswersObjectsList[i],
+            this.incidentQuestions[0][i]
+          );
+
+          if (this.operations == "edit" || this.operations == "view") {
+            this.mappingIncidentAnswers(
+              this.incidentAnswer[0].incidentAnswersObjectsList[i],
+              this.incidentQuestions[0][i].id,
+              this.incident
+            );
+            this.incidentAnswer[0].incidentAnswersObjectsList[
+              i
+            ].answer = this.getViewQuestionAnswer(
+              this.incidentQuestions[0][i],
+              this.incident.answers[i].answer
+            );
           }
+          this.incidentQuestionsEditParamTemps[i] = {};
+          this.incidentQuestionsEditParamTemps[i].parameters = [];
 
-          this.incidentQuestions[0] = this.incidentQuestionsFirst;
-
-
-          for (let i = 0; i < this.incidentQuestions[0].length; i++) {
-            this.mappingDefaultAnswers(this.incidentAnswer[0].incidentAnswersObjectsList[i], this.incidentQuestions[0][i]);
-
-            if (this.operations == 'edit' || this.operations == 'view') {
-              this.mappingIncidentAnswers(this.incidentAnswer[0].incidentAnswersObjectsList[i], this.incidentQuestions[0][i].id, this.incident);
-              this.incidentAnswer[0].incidentAnswersObjectsList[i].answer = this.getViewQuestionAnswer(this.incidentQuestions[0][i], this.incident.answers[i].answer);
-            }
-            this.incidentQuestionsEditParamTemps[i] = {};
-            this.incidentQuestionsEditParamTemps[i].parameters = [];
-
-            for (let j = 0; j < this.incidentQuestions[0][i].parametersList.length; j++) {
-              let param = {
-                "id": '',
-                "key": '',
-                "value": ''
-              };
-              this.incidentQuestionsEditParamTemps[i].parameters[j] = param;
-              this.incidentQuestionsEditParamTemps[i].parameters[j].key = this.incidentQuestions[0][i].parametersList[j].key;
-            }
-
+          for (
+            let j = 0;
+            j < this.incidentQuestions[0][i].parametersList.length;
+            j++
+          ) {
+            let param = {
+              id: "",
+              key: "",
+              value: "",
+            };
+            this.incidentQuestionsEditParamTemps[i].parameters[j] = param;
+            this.incidentQuestionsEditParamTemps[i].parameters[
+              j
+            ].key = this.incidentQuestions[0][i].parametersList[j].key;
           }
+        }
 
-          this.medicalService.getCheckupTemplate().subscribe();
-        }, async err => {
+        this.medicalService.getCheckupTemplate().subscribe();
+      },
+      async (err) => {
+        this.templateLoading = false;
 
-          this.templateLoading = false;
-
-          const alrt = await this.alrtCtrl.create({
-            header: 'Error',
-            subHeader: 'Something went wrong, can\'t load incident template',
-            buttons: ['OK']
-          });
-
-          await alrt.present();
+        const alrt = await this.alrtCtrl.create({
+          header: "Error",
+          subHeader: "Something went wrong, can't load incident template",
+          buttons: ["OK"],
         });
+
+        await alrt.present();
+      }
+    );
   }
 
   getCheckUpTemplate() {
     this.medicalService.getCheckupTemplate().subscribe(
-        val => {
-          this.templateLoading = false;
-          this.checkupTemplate = val;
-          // if(this.platform.is('cordova')){
-          //   this.checkupTemplate = JSON.parse(val.data);
-          // }
-          if (this.addCheckup) {
-            this.checkupQuestions = [];
-            this.checkupQuestions[0] = this.checkupTemplate.questionsList;
-          } else {
-            this.checkupQuestions = this.checkupTemplate.questionsList;
+      (val) => {
+        this.templateLoading = false;
+        this.checkupTemplate = val;
+        // if(this.platform.is('cordova')){
+        //   this.checkupTemplate = JSON.parse(val.data);
+        // }
+        if (this.addCheckup) {
+          this.checkupQuestions = [];
+          this.checkupQuestions[0] = this.checkupTemplate.questionsList;
+        } else {
+          this.checkupQuestions = this.checkupTemplate.questionsList;
+        }
+
+        this.checkupQuestionsFirst = this.checkupTemplate.questionsList;
+
+        for (let i = 0; i < this.checkupQuestionsFirst.length; i++) {
+          this.checkupQuestionsFirst[i].questionNumber = i;
+          this.checkupAnswer.checkupAnswersObjectsList[i] = {
+            answer: null,
+          };
+          this.checkupAnswersNoOfItems[i] = {
+            noOfItems: null,
+          };
+          this.checkupQuestionsFirst[i].editQuestion = false;
+          this.checkupQuestionsFirst[i].isEdited = false;
+        }
+        if (this.addCheckup) {
+          this.checkupQuestions[0] = this.checkupQuestionsFirst;
+        } else {
+          this.checkupQuestions = this.checkupQuestionsFirst;
+        }
+
+        if (this.addCheckup) {
+          for (let i = 0; i < this.checkupQuestions[0].length; i++) {
+            this.mappingDefaultAnswers(
+              this.checkupAnswer.checkupAnswersObjectsList[i],
+              this.checkupQuestions[0][i]
+            );
+
+            this.checkupQuestionsEditParamTemps[i] = {};
+            this.checkupQuestionsEditParamTemps[i].parameters = [];
+
+            for (
+              let j = 0;
+              j < this.checkupQuestions[0][i].parametersList.length;
+              j++
+            ) {
+              let param = {
+                id: "",
+                key: "",
+                value: "",
+              };
+              this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
+              this.checkupQuestionsEditParamTemps[i].parameters[
+                j
+              ].key = this.checkupQuestions[0][i].parametersList[j].key;
+            }
           }
+        } else {
+          for (let i = 0; i < this.checkupQuestions.length; i++) {
+            this.mappingDefaultAnswers(
+              this.checkupAnswer.checkupAnswersObjectsList[i],
+              this.checkupQuestions[i]
+            );
 
-          this.checkupQuestionsFirst = this.checkupTemplate.questionsList;
-
-          for (let i = 0; i < this.checkupQuestionsFirst.length; i++) {
-            this.checkupQuestionsFirst[i].questionNumber = i;
-            this.checkupAnswer.checkupAnswersObjectsList[i] = {
-              answer: null
-            };
-            this.checkupAnswersNoOfItems[i] = {
-              noOfItems: null
-            };
-            this.checkupQuestionsFirst[i].editQuestion = false;
-            this.checkupQuestionsFirst[i].isEdited = false;
-          }
-          if (this.addCheckup) {
-            this.checkupQuestions[0] = this.checkupQuestionsFirst;
-          } else {
-            this.checkupQuestions = this.checkupQuestionsFirst;
-          }
-
-          if (this.addCheckup) {
-            for (let i = 0; i < this.checkupQuestions[0].length; i++) {
-
-              this.mappingDefaultAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[0][i]);
-
-              this.checkupQuestionsEditParamTemps[i] = {};
-              this.checkupQuestionsEditParamTemps[i].parameters = [];
-
-              for (let j = 0; j < this.checkupQuestions[0][i].parametersList.length; j++) {
-                let param = {
-                  "id": '',
-                  "key": '',
-                  "value": ''
-                };
-                this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
-                this.checkupQuestionsEditParamTemps[i].parameters[j].key = this.checkupQuestions[0][i].parametersList[j].key;
-              }
-
+            if (this.operations == "edit" || this.operations == "view") {
+              this.mappingCheckupAnswers(
+                this.checkupAnswer.checkupAnswersObjectsList[i],
+                this.checkupQuestions[i].id,
+                this.checkup
+              );
+              this.checkupAnswer.checkupAnswersObjectsList[
+                i
+              ].answer = this.getViewQuestionAnswer(
+                this.checkupQuestions[i],
+                this.checkup.answers[i].answer
+              );
             }
 
-          } else {
-            for (let i = 0; i < this.checkupQuestions.length; i++) {
-              this.mappingDefaultAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[i]);
+            this.checkupQuestionsEditParamTemps[i] = {};
+            this.checkupQuestionsEditParamTemps[i].parameters = [];
 
-              if (this.operations == 'edit' || this.operations == 'view') {
-                this.mappingCheckupAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[i].id, this.checkup);
-                this.checkupAnswer.checkupAnswersObjectsList[i].answer =
-                    this.getViewQuestionAnswer(this.checkupQuestions[i], this.checkup.answers[i].answer);
-              }
-
-              this.checkupQuestionsEditParamTemps[i] = {};
-              this.checkupQuestionsEditParamTemps[i].parameters = [];
-
-              for (let j = 0; j < this.checkupQuestions[i].parametersList.length; j++) {
-                let param = {
-                  "id": '',
-                  "key": '',
-                  "value": ''
-                };
-                this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
-                this.checkupQuestionsEditParamTemps[i].parameters[j].key = this.checkupQuestions[i].parametersList[j].key;
-              }
-
+            for (
+              let j = 0;
+              j < this.checkupQuestions[i].parametersList.length;
+              j++
+            ) {
+              let param = {
+                id: "",
+                key: "",
+                value: "",
+              };
+              this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
+              this.checkupQuestionsEditParamTemps[i].parameters[
+                j
+              ].key = this.checkupQuestions[i].parametersList[j].key;
             }
-
           }
+        }
 
-
-          if (this.addCheckup) {
-            this.loopData = this.checkupQuestions[0];
-          } else {
-            this.loopData = this.checkupQuestions;
-          }
-
-
-        }, async err => {
-          this.templateLoading = false;
-          const alrt = await this.alrtCtrl.create({
-            header: 'Error',
-            subHeader: 'Something went wrong, can\'t load checkup template',
-            buttons: ['OK']
-          });
-          await alrt.present();
+        if (this.addCheckup) {
+          this.loopData = this.checkupQuestions[0];
+        } else {
+          this.loopData = this.checkupQuestions;
+        }
+      },
+      async (err) => {
+        this.templateLoading = false;
+        const alrt = await this.alrtCtrl.create({
+          header: "Error",
+          subHeader: "Something went wrong, can't load checkup template",
+          buttons: ["OK"],
         });
+        await alrt.present();
+      }
+    );
   }
 
   ///////////////////////////TEMPLET MEATHODS/////////////////////////////
 
   mappingDefaultAnswers(defaultcheckupAnswer, question) {
-    return defaultcheckupAnswer.answer = this.getDefaultValue(question);
+    return (defaultcheckupAnswer.answer = this.getDefaultValue(question));
   }
 
   mappingIncidentAnswers(answersObject, questionId, incident) {
     for (let i = 0; i < incident.answers.length; i++) {
       let answer = incident.answers[i];
       if (answer.questionId == questionId) {
-        return answersObject.answer = answer.answer;
+        return (answersObject.answer = answer.answer);
       }
     }
   }
@@ -677,18 +903,17 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     for (let i = 0; i < checkup.answers.length; i++) {
       let answer = checkup.answers[i];
       if (answer.questionId == questionId) {
-        return answersObject.answer = answer.answer;
+        return (answersObject.answer = answer.answer);
       }
     }
   }
 
-
   getDefaultValue(Question) {
     switch (Question.questionType.title) {
-      case 'TEXT_QUESTION':
+      case "TEXT_QUESTION":
         return "";
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_EN':
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_AR':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_EN":
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_AR":
         let val = [];
         let firstTime = true;
         let firstTextField = true;
@@ -696,27 +921,19 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         let defailtValueArray = [];
 
         for (let v = 0; v < Question.parametersList.length; v++) {
-
           if (Question.parametersList[v].key == "OPTION_HELPER_TITLE") {
-
           } else if (Question.parametersList[v].key == "OPTION_HELPER_TEXT") {
-
             defailtValueArray[counter] = {};
             defailtValueArray[counter].key = Question.parametersList[v].key;
             defailtValueArray[counter].value = "";
             counter++;
-
           } else if (Question.parametersList[v].key == "OPTION_ANSWER") {
-
             defailtValueArray[counter] = {};
             defailtValueArray[counter].key = Question.parametersList[v].key;
             defailtValueArray[counter].value = Question.parametersList[v].value;
             counter++;
-
           } else {
-
           }
-
         }
 
         for (let d = 0; d < defailtValueArray.length; d++) {
@@ -725,44 +942,41 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           } else if (defailtValueArray[d].key == "OPTION_ANSWER") {
             val[d] = defailtValueArray[d].value;
           }
-
         }
         return val;
 
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER':
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER":
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER":
         return {};
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER_WITH_EDIT':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER_WITH_EDIT":
         return {};
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_EDIT':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_EDIT":
         return [Question.parametersList[0].value];
-      case 'MULTI_SHORT_TEXT_MULTISELECT_VIEW_SELECTED':
+      case "MULTI_SHORT_TEXT_MULTISELECT_VIEW_SELECTED":
         return {};
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_TEXT_QUESTION':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_TEXT_QUESTION":
         return [Question.parametersList[0].value];
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTISELECT_ANSWER_WITH_TEXT_QUESTION':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTISELECT_ANSWER_WITH_TEXT_QUESTION":
         return [true];
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER":
         return Question.parametersList[0].value;
 
-      case 'CONSTANT_SHORT_HELPER_TEXT_QUESTION':
-      case 'CONSTANT_LONG_HELPER_TEXT_QUESTION':
-      case 'SHORT_HELPER_TEXT_QUESTION':
-      case 'LONG_HELPER_TEXT_QUESTION':
+      case "CONSTANT_SHORT_HELPER_TEXT_QUESTION":
+      case "CONSTANT_LONG_HELPER_TEXT_QUESTION":
+      case "SHORT_HELPER_TEXT_QUESTION":
+      case "LONG_HELPER_TEXT_QUESTION":
         return {};
-      case 'SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED':
+      case "SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED":
         val = [];
         firstTime = true;
         counter = 0;
         defailtValueArray = [];
 
         for (let v = 0; v < Question.parametersList.length; v++) {
-
           if (Question.parametersList[v].key == "OPTION_HELPER_TITLE") {
-
           } else if (Question.parametersList[v].key == "OPTION_HELPER_TEXT") {
             defailtValueArray[counter] = {};
             defailtValueArray[counter].key = Question.parametersList[v].key;
@@ -774,14 +988,13 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
             if (firstTime) {
               defailtValueArray[counter] = {};
               defailtValueArray[counter].key = Question.parametersList[v].key;
-              defailtValueArray[counter].value = Question.parametersList[v].value;
+              defailtValueArray[counter].value =
+                Question.parametersList[v].value;
               counter++;
               firstTime = false;
             } else {
-
             }
           }
-
         }
 
         for (let d = 0; d < defailtValueArray.length; d++) {
@@ -790,10 +1003,9 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           } else if (defailtValueArray[d].key == "OPTION_ANSWER") {
             val[d] = defailtValueArray[d].value;
           }
-
         }
         return val;
-      case 'MULTI_SHORT_TEXT_ONE_VIEW_SELECTED':
+      case "MULTI_SHORT_TEXT_ONE_VIEW_SELECTED":
         val = [];
         firstTime = true;
         firstTextField = true;
@@ -801,9 +1013,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         defailtValueArray = [];
 
         for (let v = 0; v < Question.parametersList.length; v++) {
-
           if (Question.parametersList[v].key == "OPTION_HELPER_TITLE") {
-
           } else if (Question.parametersList[v].key == "OPTION_HELPER_TEXT") {
             if (firstTextField) {
               defailtValueArray[counter] = {};
@@ -817,7 +1027,6 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
               defailtValueArray[counter].value = "00";
               counter++;
               firstTextField = true;
-
             }
 
             firstTime = true;
@@ -825,14 +1034,13 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
             if (firstTime) {
               defailtValueArray[counter] = {};
               defailtValueArray[counter].key = Question.parametersList[v].key;
-              defailtValueArray[counter].value = Question.parametersList[v].value;
+              defailtValueArray[counter].value =
+                Question.parametersList[v].value;
               counter++;
               firstTime = false;
             } else {
-
             }
           }
-
         }
 
         for (let d = 0; d < defailtValueArray.length; d++) {
@@ -841,20 +1049,17 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           } else if (defailtValueArray[d].key == "OPTION_ANSWER") {
             val[d] = defailtValueArray[d].value;
           }
-
         }
         return val;
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_AR':
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_EN':
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_AR":
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_EN":
         val = [];
         firstTime = true;
         counter = 0;
         defailtValueArray = [];
 
         for (let v = 0; v < Question.parametersList.length; v++) {
-
           if (Question.parametersList[v].key == "OPTION_HELPER_TITLE") {
-
           } else if (Question.parametersList[v].key == "OPTION_DROP_DOWN") {
             defailtValueArray[counter] = {};
             defailtValueArray[counter].key = Question.parametersList[v].key;
@@ -865,14 +1070,13 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
             if (firstTime) {
               defailtValueArray[counter] = {};
               defailtValueArray[counter].key = Question.parametersList[v].key;
-              defailtValueArray[counter].value = Question.parametersList[v].value;
+              defailtValueArray[counter].value =
+                Question.parametersList[v].value;
               counter++;
               firstTime = false;
             } else {
-
             }
           }
-
         }
 
         for (let d = 0; d < defailtValueArray.length; d++) {
@@ -881,11 +1085,10 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           } else if (defailtValueArray[d].key == "OPTION_ANSWER") {
             val[d] = defailtValueArray[d].value;
           }
-
         }
         return val;
 
-      case 'IMAGES_WITH_DESCRIPTION':
+      case "IMAGES_WITH_DESCRIPTION":
         return [];
       default:
         /*
@@ -898,18 +1101,18 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
 
   getViewQuestionAnswer(question, dbAnswer) {
     switch (question.questionType.title) {
-      case 'TEXT_QUESTION':
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER':
+      case "TEXT_QUESTION":
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return "";
         }
         return dbAnswer;
 
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER':
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER':
-      case 'LONG_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER":
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER":
+      case "LONG_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -921,7 +1124,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           value[arrayIds[i]] = true;
         }
         return value;
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER_WITH_EDIT':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTIPLE_ANSWER_WITH_EDIT":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -929,36 +1132,43 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         value = {};
         for (var i = 0; i < question.parametersList.length; i++) {
           if (question.parametersList[i].key == "OPTION_ANSWER") {
-
             if (answersList.indexOf(question.parametersList[i].value) > -1) {
-              value[i] = true
+              value[i] = true;
             } else {
               value[i] = false;
             }
           } else if (question.parametersList[i].key == "OPTION_HELPER_TEXT") {
-            if (answersList[answersList.length - 1] == null || answersList[answersList.length - 1] == "" || answersList[answersList.length - 1] == " " || answersList[answersList.length - 1] == undefined || answersList[answersList.length - 1] == "undefined") {
+            if (
+              answersList[answersList.length - 1] == null ||
+              answersList[answersList.length - 1] == "" ||
+              answersList[answersList.length - 1] == " " ||
+              answersList[answersList.length - 1] == undefined ||
+              answersList[answersList.length - 1] == "undefined"
+            ) {
               value[i] = "";
             } else {
               value[i] = answersList[answersList.length - 1];
             }
-
           }
         }
         return value;
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_EDIT':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_EDIT":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
         answersList = dbAnswer.split("$$");
         value = {};
         value[0] = answersList[0];
-        if (answersList[1] == null || answersList[1] == "" || answersList[1] == " ") {
-
+        if (
+          answersList[1] == null ||
+          answersList[1] == "" ||
+          answersList[1] == " "
+        ) {
         } else {
           value[1] = answersList[1];
         }
         return value;
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_TEXT_QUESTION':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_ONE_ANSWER_WITH_TEXT_QUESTION":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -968,7 +1178,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         value[1] = answersList[1];
 
         return value;
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTISELECT_ANSWER_WITH_TEXT_QUESTION':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_MULTISELECT_ANSWER_WITH_TEXT_QUESTION":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -980,8 +1190,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           value[1] = dbAnswer;
         }
         return value;
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_EN':
-      case 'SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_AR':
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_EN":
+      case "SHORT_TEXT_MULTISELECT_VIEW_SELECTED_NONE_ANSWER_INPUT_BOX_AR":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -991,35 +1201,35 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         arrayIds = midResult[1].split(/[$&]+/);
         value = {};
         for (let i = 0; i < arrayIds.length; i++) {
-          if ((i % 2) == 0) {
-            if (arrayValues[i] == null || arrayValues[i] == "" || arrayValues[i] == " " || arrayValues[i] == "0") {
+          if (i % 2 == 0) {
+            if (
+              arrayValues[i] == null ||
+              arrayValues[i] == "" ||
+              arrayValues[i] == " " ||
+              arrayValues[i] == "0"
+            ) {
               value[arrayIds[i]] = false;
             } else {
               value[arrayIds[i]] = true;
             }
-
           } else {
             if (arrayValues[i] == "0") {
               value[arrayIds[i]] = "";
             } else {
               value[arrayIds[i]] = arrayValues[i];
             }
-
           }
-
         }
         return value;
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_AR':
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_EN':
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_AR":
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_EN":
         let counter = 0;
         let getAnswerValueArray = [];
         let firstTimeFullArray = true;
         let questionCount = 0;
 
         for (let j = 0; j < question.parametersList.length; j++) {
-
           if (question.parametersList[j].key == "OPTION_HELPER_TITLE") {
-
           } else if (question.parametersList[j].key == "OPTION_DROP_DOWN") {
             getAnswerValueArray[counter] = {};
             getAnswerValueArray[counter].key = question.parametersList[j].key;
@@ -1033,10 +1243,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
               firstTimeFullArray = false;
               questionCount++;
             } else {
-
             }
           }
-
         }
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
@@ -1052,21 +1260,17 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           } else {
             value[i] = answersQuestionList[i];
           }
-
         }
         return value;
-      case 'SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED':
-      case 'MULTI_SHORT_TEXT_ONE_VIEW_SELECTED':
-
+      case "SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED":
+      case "MULTI_SHORT_TEXT_ONE_VIEW_SELECTED":
         counter = 0;
         getAnswerValueArray = [];
         firstTimeFullArray = true;
         questionCount = 0;
 
         for (let j = 0; j < question.parametersList.length; j++) {
-
           if (question.parametersList[j].key == "OPTION_HELPER_TITLE") {
-
           } else if (question.parametersList[j].key == "OPTION_HELPER_TEXT") {
             getAnswerValueArray[counter] = {};
             getAnswerValueArray[counter].key = question.parametersList[j].key;
@@ -1080,10 +1284,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
               firstTimeFullArray = false;
               questionCount++;
             } else {
-
             }
           }
-
         }
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
@@ -1095,22 +1297,18 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         value = {};
 
         for (let i = 0; i < getAnswerValueArray.length; i++) {
-
           value[i] = answersQuestionList[i];
-
         }
         return value;
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_AR':
-      case 'DROPDOWN_MENU_ONE_VIEW_SELECTED_EN':
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_AR":
+      case "DROPDOWN_MENU_ONE_VIEW_SELECTED_EN":
         counter = 0;
         getAnswerValueArray = [];
         firstTimeFullArray = true;
         questionCount = 0;
 
         for (let j = 0; j < question.parametersList.length; j++) {
-
           if (question.parametersList[j].key == "OPTION_HELPER_TITLE") {
-
           } else if (question.parametersList[j].key == "OPTION_DROP_DOWN") {
             getAnswerValueArray[counter] = {};
             getAnswerValueArray[counter].key = question.parametersList[j].key;
@@ -1124,10 +1322,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
               firstTimeFullArray = false;
               questionCount++;
             } else {
-
             }
           }
-
         }
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
@@ -1138,16 +1334,14 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         value = {};
 
         for (let i = 0; i < getAnswerValueArray.length; i++) {
-
           value[i] = answersQuestionList[i];
-
         }
         return value;
 
-      case 'CONSTANT_SHORT_HELPER_TEXT_QUESTION':
-      case 'CONSTANT_LONG_HELPER_TEXT_QUESTION':
-      case 'SHORT_HELPER_TEXT_QUESTION':
-      case 'LONG_HELPER_TEXT_QUESTION':
+      case "CONSTANT_SHORT_HELPER_TEXT_QUESTION":
+      case "CONSTANT_LONG_HELPER_TEXT_QUESTION":
+      case "SHORT_HELPER_TEXT_QUESTION":
+      case "LONG_HELPER_TEXT_QUESTION":
         if (dbAnswer == null || dbAnswer == "" || dbAnswer == " ") {
           return {};
         }
@@ -1157,8 +1351,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
           value[i] = answersList[i];
         }
         return value;
-      case 'IMAGES_WITH_DESCRIPTION':
-        return JSON.parse(dbAnswer)
+      case "IMAGES_WITH_DESCRIPTION":
+        return JSON.parse(dbAnswer);
     }
   }
 
@@ -1176,8 +1370,8 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     this.SINGLE_SHORT_TEXT_ONE_VIEW_SELECTED_INDEX = i;
   }
 
-  setImageOrLable(question,pramName){
-    return 'Label';
+  setImageOrLable(question, pramName) {
+    return "Label";
   }
   checkIfChangesAnswer = false;
 
@@ -1187,44 +1381,61 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
 
   ///////////////////////////////
   loading;
-  showSupportFiles:boolean;
-  async filesChange(qNumber,QID) {
+  showSupportFiles: boolean;
+  async filesChange(qNumber, QID) {
     let inputEl: HTMLInputElement = this.inputEl.nativeElement;
     let fileCount: number = inputEl.files.length;
     let faildFilesNamesSize: any[] = [];
     let faildFilesNameseExtantion: any[] = [];
-    if (fileCount > 0) { // a file was selected
+    if (fileCount > 0) {
+      // a file was selected
       for (let i = 0; i < fileCount; i++) {
         let num: number = inputEl.files.item(i).size;
         let fileName = inputEl.files.item(i).name;
         let fileExtintion: string = fileName.slice(fileName.length - 4);
-        fileExtintion = fileExtintion.replace('.', '');
-        if (num <= 26214400 && this.fileTypes.find(x => x == fileExtintion)) {
+        fileExtintion = fileExtintion.replace(".", "");
+        if (num <= 26214400 && this.fileTypes.find((x) => x == fileExtintion)) {
           let formData = new FormData();
           // debugger;
-          let file: File=inputEl.files.item(i);
+          let file: File = inputEl.files.item(i);
           let fileType = this.getFileType(inputEl.files.item(i).name);
           if (fileType == "IMAGE") {
             this.compress.compressImage(inputEl.files.item(i)).subscribe(
-                result=>{
-                  // debugger;
-                  file = result;
-                  formData.append('file', result,result.name);
-                  console.log(JSON.stringify(formData));
-                  this.organizeData(inputEl,i,formData,result,fileType,fileName,qNumber,QID);
-                },error => {
-                  console.log('😢 Oh no!', error);
-                });
-
-          }else{
+              (result) => {
+                // debugger;
+                file = result;
+                formData.append("file", result, result.name);
+                console.log(JSON.stringify(formData));
+                this.organizeData(
+                  inputEl,
+                  i,
+                  formData,
+                  result,
+                  fileType,
+                  fileName,
+                  qNumber,
+                  QID
+                );
+              },
+              (error) => {
+                console.log("😢 Oh no!", error);
+              }
+            );
+          } else {
             file = inputEl.files.item(i);
-            formData.append('file', file);
+            formData.append("file", file);
             console.log(JSON.stringify(formData));
-            this.organizeData(inputEl,i,formData,inputEl.files.item(i),fileType,fileName,qNumber,QID);
+            this.organizeData(
+              inputEl,
+              i,
+              formData,
+              inputEl.files.item(i),
+              fileType,
+              fileName,
+              qNumber,
+              QID
+            );
           }
-
-
-
         } else if (num > 26214400) {
           faildFilesNamesSize.push(inputEl.files.item(i).name);
         } else {
@@ -1232,24 +1443,46 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
         }
       }
     }
-    if (faildFilesNamesSize.length > 0 && faildFilesNameseExtantion.length <= 0) {
-      alert('Can\'t upload files name: ' + faildFilesNamesSize.join(',') + ' because it is bigger than 25 Mb');
-    } else if (faildFilesNamesSize.length <= 0 && faildFilesNameseExtantion.length > 0) {
+    if (
+      faildFilesNamesSize.length > 0 &&
+      faildFilesNameseExtantion.length <= 0
+    ) {
+      alert(
+        "Can't upload files name: " +
+          faildFilesNamesSize.join(",") +
+          " because it is bigger than 25 Mb"
+      );
+    } else if (
+      faildFilesNamesSize.length <= 0 &&
+      faildFilesNameseExtantion.length > 0
+    ) {
       this.showSupportFiles = true;
-      alert('Can\'t upload files name: ' + faildFilesNameseExtantion.join(',') + ' because it is not supported');
-    } else if (faildFilesNamesSize.length > 0 && faildFilesNameseExtantion.length > 0) {
+      alert(
+        "Can't upload files name: " +
+          faildFilesNameseExtantion.join(",") +
+          " because it is not supported"
+      );
+    } else if (
+      faildFilesNamesSize.length > 0 &&
+      faildFilesNameseExtantion.length > 0
+    ) {
       this.showSupportFiles = true;
-      alert('Can\'t upload files name: ' + faildFilesNamesSize.join(',') + ' because it is bigger than 25 Mb and' +
-          ' files name: ' + faildFilesNameseExtantion.join(',') + ' because it is not supported.');
+      alert(
+        "Can't upload files name: " +
+          faildFilesNamesSize.join(",") +
+          " because it is bigger than 25 Mb and" +
+          " files name: " +
+          faildFilesNameseExtantion.join(",") +
+          " because it is not supported."
+      );
     }
   }
 
-  organizeData(inputEl,i,formData,file,fileType,fileName,qNumber,QID){
+  organizeData(inputEl, i, formData, file, fileType, fileName, qNumber, QID) {
     // this.uploadAttach(formData);
 
     if (fileType == "IMAGE") {
-
-      this.readFile(file,qNumber,QID);
+      this.readFile(file, qNumber, QID);
     } else {
       let attach = new Postattachment();
       attach.name = fileName;
@@ -1258,12 +1491,14 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
       attach.data = file;
       attach.questionId = QID;
 
-      this.incidentAnswer[0].incidentAnswersObjectsList[qNumber].answer.push(attach);
+      this.incidentAnswer[0].incidentAnswersObjectsList[qNumber].answer.push(
+        attach
+      );
     }
   }
 
   getFileType(fileName) {
-    let pos = fileName.lastIndexOf('.');
+    let pos = fileName.lastIndexOf(".");
     let extension = fileName.substring(pos + 1);
 
     switch (extension.toLowerCase()) {
@@ -1319,10 +1554,10 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     }
   }
 
-  readFile(file: File,qNumber,QID){
+  readFile(file: File, qNumber, QID) {
     let that = this;
     let reader = new FileReader();
-    reader.onloadend = function(e){
+    reader.onloadend = function (e) {
       // you can perform an action with readed data here
       console.log(reader.result);
       let attach = new Postattachment();
@@ -1331,14 +1566,14 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
       attach.url = reader.result;
       attach.file = file;
       attach.questionId = QID;
-      that.incidentAnswer[0].incidentAnswersObjectsList[qNumber].answer.push(attach);
+      that.incidentAnswer[0].incidentAnswersObjectsList[qNumber].answer.push(
+        attach
+      );
     };
     reader.readAsDataURL(file);
   }
 
-
-
-  setSelectedClassFromEdit(data){
+  setSelectedClassFromEdit(data) {
     let item = new Class();
     item.id = data.id;
     item.name = data.name;
@@ -1351,7 +1586,7 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     this.selectedClass = item;
   }
 
-  setSelectedStudentFromEdit(value){
+  setSelectedStudentFromEdit(value) {
     let students = new Student();
     students.classes.id = value.classes.id;
     students.classes.name = value.classes.name;
@@ -1363,10 +1598,11 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     students.id = value.id;
     students.name = value.name;
     students.address = value.address;
-    students.searchByClassGrade = value.classes.grade.name + " - " + value.classes.name;
+    students.searchByClassGrade =
+      value.classes.grade.name + " - " + value.classes.name;
     this.selectedStudent = students;
   }
-  setIncidentTempletFromEdit(val){
+  setIncidentTempletFromEdit(val) {
     this.templateLoading = false;
     this.incidentTemplate = val;
 
@@ -1377,10 +1613,10 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     for (let i = 0; i < this.incidentQuestionsFirst.length; i++) {
       this.incidentQuestionsFirst[i].questionNumber = i;
       this.incidentAnswer[0].incidentAnswersObjectsList[i] = {
-        answer: null
+        answer: null,
       };
       this.incidentAnswersNoOfItems[i] = {
-        noOfItems: null
+        noOfItems: null,
       };
       this.incidentQuestionsFirst[i].editQuestion = false;
       this.incidentQuestionsFirst[i].isEdited = false;
@@ -1388,32 +1624,47 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
 
     this.incidentQuestions[0] = this.incidentQuestionsFirst;
 
-
     for (let i = 0; i < this.incidentQuestions[0].length; i++) {
-      this.mappingDefaultAnswers(this.incidentAnswer[0].incidentAnswersObjectsList[i], this.incidentQuestions[0][i]);
+      this.mappingDefaultAnswers(
+        this.incidentAnswer[0].incidentAnswersObjectsList[i],
+        this.incidentQuestions[0][i]
+      );
 
-      if (this.operations == 'edit' || this.operations == 'view') {
-        this.mappingIncidentAnswers(this.incidentAnswer[0].incidentAnswersObjectsList[i], this.incidentQuestions[0][i].id, this.incident);
-        this.incidentAnswer[0].incidentAnswersObjectsList[i].answer = this.getViewQuestionAnswer(this.incidentQuestions[0][i], this.incident.answers[i].answer);
+      if (this.operations == "edit" || this.operations == "view") {
+        this.mappingIncidentAnswers(
+          this.incidentAnswer[0].incidentAnswersObjectsList[i],
+          this.incidentQuestions[0][i].id,
+          this.incident
+        );
+        this.incidentAnswer[0].incidentAnswersObjectsList[
+          i
+        ].answer = this.getViewQuestionAnswer(
+          this.incidentQuestions[0][i],
+          this.incident.answers[i].answer
+        );
       }
       this.incidentQuestionsEditParamTemps[i] = {};
       this.incidentQuestionsEditParamTemps[i].parameters = [];
 
-      for (let j = 0; j < this.incidentQuestions[0][i].parametersList.length; j++) {
+      for (
+        let j = 0;
+        j < this.incidentQuestions[0][i].parametersList.length;
+        j++
+      ) {
         let param = {
-          "id": '',
-          "key": '',
-          "value": ''
+          id: "",
+          key: "",
+          value: "",
         };
         this.incidentQuestionsEditParamTemps[i].parameters[j] = param;
-        this.incidentQuestionsEditParamTemps[i].parameters[j].key = this.incidentQuestions[0][i].parametersList[j].key;
+        this.incidentQuestionsEditParamTemps[i].parameters[
+          j
+        ].key = this.incidentQuestions[0][i].parametersList[j].key;
       }
-
     }
   }
 
-
-  setCheckupTempletFromEdit(val){
+  setCheckupTempletFromEdit(val) {
     this.templateLoading = false;
     this.checkupTemplate = val;
     if (this.addCheckup) {
@@ -1428,10 +1679,10 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
     for (let i = 0; i < this.checkupQuestionsFirst.length; i++) {
       this.checkupQuestionsFirst[i].questionNumber = i;
       this.checkupAnswer.checkupAnswersObjectsList[i] = {
-        answer: null
+        answer: null,
       };
       this.checkupAnswersNoOfItems[i] = {
-        noOfItems: null
+        noOfItems: null,
       };
       this.checkupQuestionsFirst[i].editQuestion = false;
       this.checkupQuestionsFirst[i].isEdited = false;
@@ -1444,51 +1695,71 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
 
     if (this.addCheckup) {
       for (let i = 0; i < this.checkupQuestions[0].length; i++) {
-
-        this.mappingDefaultAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[0][i]);
+        this.mappingDefaultAnswers(
+          this.checkupAnswer.checkupAnswersObjectsList[i],
+          this.checkupQuestions[0][i]
+        );
 
         this.checkupQuestionsEditParamTemps[i] = {};
         this.checkupQuestionsEditParamTemps[i].parameters = [];
 
-        for (let j = 0; j < this.checkupQuestions[0][i].parametersList.length; j++) {
+        for (
+          let j = 0;
+          j < this.checkupQuestions[0][i].parametersList.length;
+          j++
+        ) {
           let param = {
-            "id": '',
-            "key": '',
-            "value": ''
+            id: "",
+            key: "",
+            value: "",
           };
           this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
-          this.checkupQuestionsEditParamTemps[i].parameters[j].key = this.checkupQuestions[0][i].parametersList[j].key;
+          this.checkupQuestionsEditParamTemps[i].parameters[
+            j
+          ].key = this.checkupQuestions[0][i].parametersList[j].key;
         }
-
       }
-
     } else {
       for (let i = 0; i < this.checkupQuestions.length; i++) {
-        this.mappingDefaultAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[i]);
+        this.mappingDefaultAnswers(
+          this.checkupAnswer.checkupAnswersObjectsList[i],
+          this.checkupQuestions[i]
+        );
 
-        if (this.operations == 'edit' || this.operations == 'view') {
-          this.mappingCheckupAnswers(this.checkupAnswer.checkupAnswersObjectsList[i], this.checkupQuestions[i].id, this.checkup);
-          this.checkupAnswer.checkupAnswersObjectsList[i].answer =
-              this.getViewQuestionAnswer(this.checkupQuestions[i], this.checkup.answers[i].answer);
+        if (this.operations == "edit" || this.operations == "view") {
+          this.mappingCheckupAnswers(
+            this.checkupAnswer.checkupAnswersObjectsList[i],
+            this.checkupQuestions[i].id,
+            this.checkup
+          );
+          this.checkupAnswer.checkupAnswersObjectsList[
+            i
+          ].answer = this.getViewQuestionAnswer(
+            this.checkupQuestions[i],
+            this.checkup.answers[i].answer
+          );
         }
 
         this.checkupQuestionsEditParamTemps[i] = {};
         this.checkupQuestionsEditParamTemps[i].parameters = [];
 
-        for (let j = 0; j < this.checkupQuestions[i].parametersList.length; j++) {
+        for (
+          let j = 0;
+          j < this.checkupQuestions[i].parametersList.length;
+          j++
+        ) {
           let param = {
-            "id": '',
-            "key": '',
-            "value": ''
+            id: "",
+            key: "",
+            value: "",
           };
           this.checkupQuestionsEditParamTemps[i].parameters[j] = param;
-          this.checkupQuestionsEditParamTemps[i].parameters[j].key = this.checkupQuestions[i].parametersList[j].key;
+          this.checkupQuestionsEditParamTemps[i].parameters[
+            j
+          ].key = this.checkupQuestions[i].parametersList[j].key;
         }
-
       }
-
     }
-
 
     if (this.addCheckup) {
       this.loopData = this.checkupQuestions[0];
@@ -1496,7 +1767,4 @@ export class MedicalCareNewMedicalReportPage implements OnInit {
       this.loopData = this.checkupQuestions;
     }
   }
-
-
-
 }
