@@ -1,20 +1,19 @@
 /* tslint:disable:max-line-length */
-import { Injectable } from '@angular/core';
-import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {Url_domain} from '../../models/url_domain';
+import { Injectable } from "@angular/core";
+import { HttpClient, HttpHeaders } from "@angular/common/http";
+import { Url_domain } from "../../models/url_domain";
 // import 'rxjs/add/observable/fromPromise';
 // import 'rxjs/add/operator/map';
 // import 'rxjs/add/operator/mergeMap';
-import {Custom_reports_data} from '../../models/custom_reports_data';
-import {BehaviorSubject, from} from 'rxjs';
-import {HTTP} from '@ionic-native/http/ngx';
-import {Platform} from '@ionic/angular';
+import { Custom_reports_data } from "../../models/custom_reports_data";
+import { BehaviorSubject, from } from "rxjs";
+import { HTTP } from "@ionic-native/http/ngx";
+import { Platform } from "@ionic/angular";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root",
 })
 export class AccountService {
-
   DomainUrl: Url_domain;
   private value: any = [];
   private accountFeature: any = [];
@@ -38,48 +37,66 @@ export class AccountService {
   private _reportId: number;
   private _userId: number;
 
-  private _user:any;
+  private _user: any;
 
   menuFeatures: BehaviorSubject<object> = new BehaviorSubject(null);
 
-  constructor(private http: HttpClient,private httpN:HTTP, private platform:Platform) {
+  constructor(
+    private http: HttpClient,
+    private httpN: HTTP,
+    private platform: Platform
+  ) {
     this.DomainUrl = new Url_domain();
   }
 
-  getAccountRoles(subHeader: string) {
+  getAccountLogoUrl() {
+    return (
+      this.DomainUrl.Domain +
+      "/admissionApplication.ent/getImage.ent?url=https://storage.googleapis.com/entrepreware-education.appspot.com/schoolsLogos/" +
+      this.userAccount.logo
+    );
+  }
 
-    const httpOptions =  {headers: new HttpHeaders({
+  getAccountRoles(subHeader: string) {
+    const httpOptions = {
+      headers: new HttpHeaders({
         // 'Access-Control-Allow-Origin' : 'http://localhost:8100',
         // 'Access-Control-Allow-Methods': 'POST, GET, OPTIONS, PUT',
         // 'Accept':'application/json',
         // 'content-type':'application/json',
-        Authorization : subHeader
-      })};
+        Authorization: subHeader,
+      }),
+    };
     // if(this.platform.is('cordova')){
     //   return from(this.httpN.post(this.DomainUrl.Domain + '/authentication/authenticator.ent?operationId=3', {}, {'Authorization' : subHeader}));
     // }else {
-      return this.http.post(this.DomainUrl.Domain + '/authentication/authenticator.ent?operationId=3', null);
+    return this.http.post(
+      this.DomainUrl.Domain + "/authentication/authenticator.ent?operationId=3",
+      null
+    );
     // }
   }
 
-
   getCustomReports(subHeader: string) {
-    const httpOptions =  {headers: new HttpHeaders({
+    const httpOptions = {
+      headers: new HttpHeaders({
         // 'Access-Control-Allow-Origin' : 'http://localhost:8100',
         // 'Access-Control-Allow-Methods': 'POST, GET, OPTIONS, PUT',
         // 'Accept':'application/json',
         // 'content-type':'application/json',
-        Authorization : subHeader
-      })};
+        Authorization: subHeader,
+      }),
+    };
     // if(this.platform.is('cordova')){
     //   return from(this.httpN.get(this.DomainUrl.Domain + '/authentication/report.ent/accountReports.ent', {}, {'Authorization' : subHeader}));
     // }else{
-      return this.http.get(this.DomainUrl.Domain + '/authentication/report.ent/accountReports.ent');
+    return this.http.get(
+      this.DomainUrl.Domain + "/authentication/report.ent/accountReports.ent"
+    );
     // }
   }
 
   setDate(data: any) {
-
     this.value = data;
     if (this.value.data != null) {
       this.value = JSON.parse(this.value.data);
@@ -103,7 +120,6 @@ export class AccountService {
     this.menuFeatures.next(this.value.accountFeatures);
   }
 
-
   setCustomReport(data: any) {
     this.customReportsList = [];
     this.customReportsValue = data;
@@ -116,15 +132,16 @@ export class AccountService {
     }
   }
 
-
   getTags(subHeader: string) {
-    const httpOptions =  {headers: new HttpHeaders({
+    const httpOptions = {
+      headers: new HttpHeaders({
         // 'Access-Control-Allow-Origin' : 'http://localhost:8100',
         // 'Access-Control-Allow-Methods': 'POST, GET, OPTIONS, PUT',
         // 'Accept':'application/json',
         // 'content-type':'application/json',
-        Authorization : subHeader
-      })};
+        Authorization: subHeader,
+      }),
+    };
 
     // if(this.platform.is('cordova')){
     //   from(this.httpN.get(this.DomainUrl.Domain + '/authentication/tag.ent?branchesIds=' + this.accountBranchesListIds, {},{'Authorization' : subHeader})).subscribe(value => {
@@ -135,7 +152,13 @@ export class AccountService {
     //   });
     // }else {
 
-      this.http.get(this.DomainUrl.Domain + '/authentication/tag.ent?branchesIds=' + this.accountBranchesListIds).subscribe(value => {
+    this.http
+      .get(
+        this.DomainUrl.Domain +
+          "/authentication/tag.ent?branchesIds=" +
+          this.accountBranchesListIds
+      )
+      .subscribe((value) => {
         this.Arry = value;
         for (const tag of this.Arry) {
           this._tagArry.push(tag);
@@ -187,7 +210,6 @@ export class AccountService {
     this._accountBranchesListIds = value;
   }
 
-
   get tagArry(): any {
     return this._tagArry;
   }
@@ -196,7 +218,6 @@ export class AccountService {
     this._tagArry = value;
   }
 
-
   get reportPage(): string {
     return this._reportPage;
   }
@@ -204,7 +225,6 @@ export class AccountService {
   set reportPage(value: string) {
     this._reportPage = value;
   }
-
 
   get reportId(): any {
     return this._reportId;
@@ -230,11 +250,9 @@ export class AccountService {
     this._userBranchId = value;
   }
 
-
   get userAccount(): any {
     return this._userAccount;
   }
-
 
   get user(): any {
     return this._user;
